204计数质数埃筛法优化做法，修改类型为long long防止溢出

+++ leetcode/editor/cn/count-primes.cpp
@@ -47,13 +47,13 @@ public:
     int countPrimes(int n)
     {
         vector<bool> isPrime(n, true);
-        int count = 0;
-        for (int i = 2; i < n; i++)
+        long long count = 0;
+        for (long long i = 2; i < n; i++)
         {
             if (isPrime[i])
             {
                 count++;
-                for (int j = i * i; j < n; j += i)
+                for (long long j = i * i; j < n; j += i)
                 {
                     isPrime[j] = false;
                 }
